requirements.txt
fastapi
gunicorn
uvicorn
azure-core==1.32.0
azure-identity==1.19.0
openai
requests
python-dotenv
requests>=2.31.0
langchain
chainlit
pydantic==2.10.1